Add StardustDID::from_block
